+++ src/Ampersand/Options/ProtoOptsParser.hs
@@ -5,19 +5,24 @@ import Ampersand.Misc.Defaults (defaultDirPrototype)
 import Ampersand.Misc.HasClasses
 import Ampersand.Options.FSpecGenOptsParser
 import Ampersand.Options.Utils
+import Data.List.Split (splitWhen)
 import Options.Applicative
 import Options.Applicative.Builder.Extra
+import RIO.Char ( toLower )
 
 protoOptsParser :: Parser ProtoOpts
 protoOptsParser =
   standardToProtoType <$>
     (  ProtoOpts 
-         <$> outputLanguageP
+         <$> forceReinstallP
+         <*> outputLanguageP
          <*> fSpecGenOptsParser False
          <*> optional dirPrototypeP
+         <*> optional dirCustomizationsP
+         <*> zwolleVersionP
          <*> generateFrontendP
          <*> generateBackendP
-         <*> checkCompilerVersionP
+         <*> frontendVersionP
          <*> generateMetamodelP
     )
 
@@ -29,6 +34,16 @@ protoOptsParser =
         Standard -> set recipeL Prototype opts
         _ -> opts
 
+    forceReinstallP :: Parser Bool
+    forceReinstallP =
+      switch
+        ( long "force-reinstall-framework"
+            <> help
+              ( "Re-install the prototype framework. This discards any previously "
+                  <> "installed version."
+              )
+        )
+
     dirPrototypeP :: Parser String
     dirPrototypeP =
       strOption
@@ -39,13 +54,47 @@ protoOptsParser =
             <> help "Specify the directory where the prototype will be generated"
         )
 
+    dirCustomizationsP :: Parser [String]
+    dirCustomizationsP =
+      splitWhen (== ';')
+        <$> strOption
+          ( long "customizations"
+              <> metavar "DIR;DIR;.."
+              <> help "Copy one or more directories into the generated prototype. "
+          )
+
+    zwolleVersionP :: Parser String
+    zwolleVersionP =
+      strOption
+        ( long "prototype-framework-version"
+            <> metavar "VERSION"
+            <> value "v1.6.0"
+            <> showDefault
+            <> help
+              ( "Tag, branch or SHA of the prototype framework on Github. "
+                  <> "Normally you shouldn't need to use anohter version "
+                  <> "than the default. Only a developer of the framework "
+                  <> "can make good use of it. "
+              )
+        )
+    frontendVersionP :: Parser FrontendVersion
+    frontendVersionP = toFrontendVersion <$> strOption
+            ( long "frontend-version"
+            <> metavar "VERSION"
+            <> value "angularjs"
+            <> showDefault
+            <> help  "Temporary switch to enable the development of the new angular frontend."
+            )
+        where
+          toFrontendVersion :: String -> FrontendVersion
+          toFrontendVersion x = case map toLower x of
+                                    "angular"   -> Angular 
+                                    "angularjs" -> AngularJS
+                                    _           -> AngularJS    
     generateFrontendP :: Parser Bool
-    generateFrontendP =
-      boolFlags
-        True
-        "frontend"
-        "Generate prototype frontend files (Angular application)"
-        mempty
+    generateFrontendP = boolFlags True "frontend"
+            "Generate prototype frontend files (Angular application)"
+            mempty
 
     generateBackendP :: Parser Bool
     generateBackendP =
@@ -53,14 +102,6 @@ protoOptsParser =
         True
         "backend"
         "Generate backend files (PHP application)"
-        mempty
-
-    checkCompilerVersionP :: Parser Bool
-    checkCompilerVersionP = 
-      boolFlags 
-        True
-        "check-compiler-version"
-        "Check compiler version constraints set by prototype framework (backend)"
         mempty
 
     -- This metamodel shows what the meatgrinder has made. This is useful for building prototypes that build on the meatgrinder.