+++ src/Ampersand/Test/Regression.hs
@@ -11,6 +11,7 @@ import           Conduit
 import           Data.Yaml
 import           RIO.Process
 import qualified RIO.ByteString.Lazy as BL
+import qualified RIO.ByteString.Lazy.Partial as BLPartial
 import qualified RIO.Text as T
 import           System.Directory
 import           System.FilePath
@@ -219,8 +220,15 @@ testAdlfile indent dir adl tinfo = do
                                          else "ExitFailure "<>display (exitcode tinfo)
                                        )
                      <>", Actual: "<>display (tshow exit_code)<>")"
-          mapM_ (logWarn  . indnt) . toUtf8Builders $ out
+          mapM_ (logError  . indnt) . toUtf8Builders $ out
+           
+          logError . indnt $ "------------------- " 
           mapM_ (logError . indnt) . toUtf8Builders $ err
+          logError . indnt . display $ "out: " <> tshow (nrOfLines out) <> " lines."
+          logError . indnt . display $ "err: " <> tshow (nrOfLines err) <> " lines."
+          logError  "❗❗❗ -------------------" 
+        where
+          nrOfLines = length . BL.split (BLPartial.head "\n")
      indnt :: Utf8Builder -> Utf8Builder
      indnt = ("    " <>)
      toUtf8Builders :: BL.ByteString -> [Utf8Builder]