+++ ReleaseNotes.md
@@ -1,6 +1,8 @@
 # Release notes of Ampersand
 
 ## Unreleased changes
+
+* [Issue #prototype/69 ](https://github.com/AmpersandTarski/prototype/issues/69 )Auxiliary work to facilitate the implementation of the new Angular frontend. 
 * [Issue #425](https://github.com/AmpersandTarski/Ampersand/issues/425) Add check on release notes for every pull request
 * [Issue #941](https://github.com/AmpersandTarski/Ampersand/issues/941) Fix and document purpose and difference of attNull and attDBNull aspects of SqlAttribute
 * [Issue #1093](https://github.com/AmpersandTarski/Ampersand/issues/1093) Don't download prototype framework from github anymore. The framework must be deployed first via another method (manual, docker or composer+packagist)